Fix write-region and its subroutines for buffers > 2GB.
 src/fileio.c (a_write, e_write): Modify declaration of arguments and
 local variables to support buffers larger than 2GB.
 (Fcopy_file): Use EMACS_INT for return value of emacs_read.
 src/sysdep.c (emacs_write, emacs_read): Use ssize_t for last
 argument, local variables, and return value.
 src/lisp.h: Update prototypes of emacs_write and emacs_read.
 src/sound.c (vox_write): Use ssize_t for return value of emacs_write.

+++ src/lisp.h
@@ -3346,8 +3346,8 @@ extern long get_random (void);
 extern void seed_random (long);
 extern int emacs_open (const char *, int, int);
 extern int emacs_close (int);
-extern int emacs_read (int, char *, unsigned int);
-extern int emacs_write (int, const char *, unsigned int);
+extern ssize_t emacs_read (int, char *, ssize_t);
+extern ssize_t emacs_write (int, const char *, ssize_t);
 enum { READLINK_BUFSIZE = 1024 };
 extern char *emacs_readlink (const char *, char [READLINK_BUFSIZE]);
 #ifndef HAVE_MEMSET
